More bug fixes
Added weight to Bishops and Knights to prioritize them over Pawns and fixed a bug where the AI gave up when it predicted loss. As of today it has won 12 out of 16 games against classmate AI, with two of those losses being against evenly matched AI that my program won against in a rematch.